Updated vars.OngoingFight logic
Checking if current.Magic has changed doesn't work soon enough with this particular pointer (it worked with a prior one I had tested).

Instead, we'll check the absolute difference in IGT from frame to frame during a fight. IGT is set to the new value (either on continue or loading a save) quickly enough to be more valuable.