Autostart for NG+ runs
I started by adding a setting for NG+ that, when enabled, would check for starts less leniently. But that allows some file loads from the main menu to false start, and although those false starts could be immediately caught and auto-reset, that's still not ideal.

This solution should work perfectly, in combination with the refresh rate being set to 30. With refresh at 60, we would catch a false stop in IGT every second, probably because the game was programmed for 30.