Fix: False splits in boss fights
The enemy count (very rarely) flickers to 0 for less than a single frame. This may be a quirk of the game's conversion from 30 FPS, but at any rate, it's impossible to predict when it will occur.

This fix checks two separate enemy count values in memory for agreement before splitting. In hours of testing, I haven't seen two values flicker simultaneously (I checked 22 pointers concurrently, actually), as it's a rare phenomenon to begin with. There are two or three other regions in memory that can be piled onto this check, but that's almost certainly unnecessary.